Reduce Mapit auto-scaling groups from 8 to 1
This commit reduces the number of Mapit auto-scaling groups from 8 to 1. Previously, each auto-scaling group was responsible for bringing up a particular-numbered instance, which was replicated for each of the instances required. This defeats the purpose of an auto-scaling group to maintain a cluster of healthy instances, and can instead be merged into a single auto-scaling group to achieve the same end goal.