Replace WAF Classic with WAF v2
WAF v2 is the latest version of the AWS WAF, replacing WAF Classic in
November 2019.

Upgrading brings us some benefits [1] but none are included in this commit.
Importantly for our use case, WAF v2 permits matching requests with rules
using the FORWARDED_IP rather than the IP. This is required for rate limiting
since we would like to rate limit users based on the True-Client-IP (the
real IP address of the user) rather than using the IP addresses of Fastly.

[1]: https://docs.aws.amazon.com/waf/latest/developerguide/waf-migrating-why-migrate.html