Create a shared memcached cluster for Mapit
infra-security-groups will need to be applied first, as app-mapit
depends on a new output.

The r6g instance type is the default if you click through in the AWS
console. AWS suggest that the ARM based processers work well with
elasticache:

https://aws.amazon.com/about-aws/whats-new/2020/10/amazon-elasticache-now-supports-m6g-and-r6g-graviton2-based-instances/

The choice between m6g and r6g is arbitrary - r6g comes first in the
dropdown in the console.

The choice of `large` (rather than `xlarge` or higher) is also
arbitrary. I'm assuming that the failure mode for this will be graceful,
as it's just a cache, so starting small and scaling based on metrics
feels reasonable. You could argue the other way though (start big and
make it smaller based on metrics).